regen imdb seed
