+++ seed/ruby-sdk-v2/imdb/lib/seed/client.rb
@@ -1,21 +1,21 @@
 
 
 module Seed
-    class Client
-        # @return [Seed::Client]
-        def initialize(base_url:, token:)
-            @raw_client = Seed::Internal::Http::RawClient.new(
-                base_url: base_url,
-                headers: {
-                    'User-Agent':'fern_imdb/0.0.1',
-                    'X-Fern-Language':'Ruby',
-                    Authorization:"Bearer #{token}"
-                }
-            )
-        end
-        # @return [Seed::Imdb::Client]
-        def imdb
-            @imdb ||= Seed::Imdb::Client.new(client: @raw_client)
-        end
+  class Client
+    # @return [Seed::Client]
+    def initialize(base_url:, token:)
+      @raw_client = Seed::Internal::Http::RawClient.new(
+        base_url: base_url,
+        headers: {
+          'User-Agent':'fern_imdb/0.0.1',
+          'X-Fern-Language':'Ruby',
+          Authorization:"Bearer #{token}"
+        }
+      )
     end
+    # @return [Seed::Imdb::Client]
+    def imdb
+      @imdb ||= Seed::Imdb::Client.new(client: @raw_client)
+    end
+  end
 end
