--- seed/ruby-sdk-v2/imdb/lib/seed/imdb/create_movie_request.rb
@@ -1,10 +1,0 @@
-# frozen_string_literal: true
-
-module Seed
-    module Types
-        class CreateMovieRequest < Internal::Types::Model
-            field :title, String, optional: false, nullable: false
-            field :rating, Integer, optional: false, nullable: false
-        end
-    end
-end
