--- seed/ruby-sdk-v2/imdb/lib/seed/imdb/movie.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-module Seed
-    module Types
-        class Movie < Internal::Types::Model
-            field :id, String, optional: false, nullable: false
-            field :title, String, optional: false, nullable: false
-            field :rating, Integer, optional: false, nullable: false
-        end
-    end
-end
