+++ seed/ruby-sdk-v2/imdb/lib/seed/imdb/types/create_movie_request.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module Seed
+  module Imdb
+    module Types
+      class CreateMovieRequest < Internal::Types::Model
+        field :title, String, optional: false, nullable: false
+        field :rating, Integer, optional: false, nullable: false
+
+      end
+    end
+  end
+end
