+++ seed/ruby-sdk-v2/imdb/lib/seed/imdb/types/movie.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+module Seed
+  module Imdb
+    module Types
+      class Movie < Internal::Types::Model
+        field :id, String, optional: false, nullable: false
+        field :title, String, optional: false, nullable: false
+        field :rating, Integer, optional: false, nullable: false
+
+      end
+    end
+  end
+end
