+++ seed/ruby-sdk-v2/imdb/lib/seed/internal/errors/type_error.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-module seed
+module Seed
   module Internal
     module Errors
       class TypeError < StandardError
