+++ seed/ruby-sdk-v2/imdb/lib/seed/internal/http/base_request.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-module seed
+module Seed
   module Internal
     module Http
       # @api private
@@ -12,7 +12,7 @@ module seed
         # @param method [String] The HTTP method for the request (:get, :post, etc.)
         # @param headers [Hash] Additional headers for the request (optional)
         # @param query [Hash] Query parameters for the request (optional)
-        # @param request_options [seed::RequestOptions, Hash{Symbol=>Object}, nil]
+        # @param request_options [Seed::RequestOptions, Hash{Symbol=>Object}, nil]
         def initialize(base_url:, path:, method:, headers: {}, query: {}, request_options: {})
           @base_url = base_url
           @path = path
