+++ seed/ruby-sdk-v2/imdb/lib/seed/internal/json/serializable.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 
-module seed
+module Seed
   module Internal
     module JSON
-      class Serializable
+      module Serializable
         # Loads data from JSON into its deserialized form
         #
         # @param str [String] Raw JSON to load into an object
