+++ seed/ruby-sdk-v2/imdb/lib/seed/internal/multipart/multipart_encoder.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-module seed
+module Seed
   module Internal
     module Multipart
       # Encodes parameters into a `multipart/form-data` payload as described by RFC
