+++ seed/ruby-sdk-v2/imdb/lib/seed/internal/multipart/multipart_form_data.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-module seed
+module Seed
   module Internal
     module Multipart
       # @api private
