+++ seed/ruby-sdk-v2/imdb/lib/seed/internal/multipart/multipart_form_data_part.rb
@@ -2,7 +2,7 @@
 
 require "securerandom"
 
-module seed
+module Seed
   module Internal
     module Multipart
       # @api private
