+++ seed/ruby-sdk-v2/imdb/lib/seed/internal/types/boolean.rb
@@ -1,10 +1,10 @@
 # frozen_string_literal: true
 
-module seed
+module Seed
   module Internal
     module Types
       module Boolean
-      extend seed::Internal::Types::Union
+        include Seed::Internal::Types::Union
 
         member TrueClass
         member FalseClass
