+++ seed/ruby-sdk-v2/imdb/lib/seed/internal/types/model/field.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
-module seed
+module Seed
   module Internal
     module Types
       class Model
