+++ seed/ruby-sdk-v2/imdb/lib/seed/internal/types/utils.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-module seed
+module Seed
   module Internal
     module Types
       # Utilities for dealing with and checking types
